feat(modules): start moving to new modules format
