+++ packages/generator/src/connection/__tests__/ConnectionGenerator.spec.js
@@ -8,7 +8,7 @@ import {
   getFixturePath,
 } from '../../../test/helpers';
 
-import { getConfigDir } from '../../utils';
+import { getConfigDir } from '../../config';
 
 const connectionGenerator = path.join(__dirname, '..');
