+++ packages/generator/src/graphqlrc.json
@@ -1,14 +1,14 @@
 {
   "directories": {
-    "source": "src",
-    "connection": "connection",
-    "loader": "loader",
-    "model": "model",
+    "source": "src/modules",
+    "connection": "",
+    "loader": "",
+    "model": "",
     "mutation": "mutation",
-    "mutation_test": "mutation/__tests__",
-    "type": "type",
-    "type_test": "type/__tests__",
-    "interface": "interface"
+    "mutation_test": "mutation",
+    "type": "",
+    "type_test": "",
+    "interface": "../../../graphql/interface"
   },
   "files": {
     "schema": "schema"
