+++ packages/generator/src/loader/__tests__/LoaderGenerator.spec.js
@@ -7,7 +7,7 @@ import {
   getFileContent,
   getFixturePath,
 } from '../../../test/helpers';
-import { getConfigDir } from '../../utils';
+import { getConfigDir } from '../../config';
 
 const loaderGenerator = path.join(__dirname, '..');
