+++ packages/generator/src/loader/templates/LoaderWithSchema.js.template
@@ -1,9 +1,8 @@
 // @flow
-
 import DataLoader from 'dataloader';
-import <%= name %>Model from '<%= directories.model %>/<%= name %>';
-import connectionFromMongoCursor from './ConnectionFromMongoCursor';
-import mongooseLoader from './mongooseLoader';
+import <%= name %>Model from './<%= name %>Model';
+import { connectionFromMongoCursor, mongooseLoader } from '@entria/graphql-mongoose-loader';
+import type { GraphQLContext } from '../../TypeDefinition';
 
 type <%= name %>Type = {
   id: string,
@@ -31,26 +30,27 @@ export default class <%= name %> {
 
 export const getLoader = () => new DataLoader(ids => mongooseLoader(<%= name %>Model, ids));
 
-const viewerCanSee = (viewer, data) => {
-  // TODO: handle security
-  return true;
-};
+const viewerCanSee = () => true;
 
-export const load = async ({ user: viewer, dataloaders }, id) => {
+export const load = async ({ user: viewer, dataloaders }: GraphQLContext, id: string) => {
   if (!id) return null;
 
-  const data = await dataloaders.<%= rawName %>Loader.load(id.toString());
+  try {
+    const data = await dataloaders.<%= rawName %>Loader.load(id.toString());
 
-  if (!data) return null;
+    if (!data) return null;
 
-  return viewerCanSee(viewer, data) ? new <%= name %>(data) : null;
+    return viewerCanSee() ? new <%= name %>(data) : null;
+  } catch (err) {
+    return null
+  };
 };
 
-export const clearCache = ({ dataloaders }, id) => {
+export const clearCache = ({ dataloaders }: GraphQLContext, id: string) => {
   return dataloaders.<%= rawName %>Loader.clear(id.toString());
 };
 
-export const load<%= pluralName %> = async (context, args) => {
+export const load<%= pluralName %> = async (context: GraphQLContext, args: ConnectionArguments) => {
   // TODO: specify conditions
   const <%= pluralCamelCaseName %> = <%= name %>Model.find({});
